requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==21.3.0
certifi==2021.10.8
cffi==1.15.0
click==8.1.3
colorama==0.4.6
cryptography==36.0.1
Django==3.2.7
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
h11==0.12.0
idna==3.3
joblib==1.2.0
markdown2==2.4.1
nltk==3.8.1
outcome==1.1.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==21.0.0
pytz==2021.1
regex==2022.10.31
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tqdm==4.64.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wsproto==1.0.0
